Initial Database ABC, ABCMetaEnforcedAttrs, ClassIdentifier, tests.
Defines initial Database API with an abstract base class, required
methods/attrs.

Metaclass ABCMetaEnforcedAttrs(abc.ABCMeta) that enforces existence of
attr names listed in required_attributes list of strings in a base or
metaclass that uses it as a metaclass.
Database uses this to enforce default_avatar_path in subclasses.

Define ClassIdentifier class inheriting from NamedTuple that holds .id
and .name for a class - allowing the id to be anything the backend needs
to use, with the string name of the class. This avoids difficulty with
supporting existing JSON database, as well as a uniform API between
backends, and with a db index will not be costly.

Signed-off-by: toonarmycaptain <toonarmycaptain@hotmail.com>